// frontend/src/__generated__/TodoListRelayContainerUpdateMutation.graphql.ts
export type PriorityEnum = "HIGH" | "LOW" | "MEDIUM" | "%future added value";
export type TodoUpdateInput = {
  priority?: PriorityEnum | null | undefined;
  progress?: number | null | undefined;
  todoTitle?: string | null | undefined;
};
export type TodoListRelayContainerUpdateMutation$variables = {
  id: string;
  input: TodoUpdateInput;
};
export type TodoListRelayContainerUpdateMutation$data = {
  readonly updateTodo: {
    readonly __typename: "TodoType";
    readonly id: string;
    readonly priority: PriorityEnum;
    readonly progress: number;
    readonly todoTitle: string;
  } | {
    readonly __typename: "ValidationError";
    readonly message: string;
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
};
export type TodoListRelayContainerUpdateMutation = {
  response: TodoListRelayContainerUpdateMutation$data;
  variables: TodoListRelayContainerUpdateMutation$variables;
};